pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "conflict-duplication-detector"
version = "0.1.0"
description = "Detect conflicts, duplications, and inconsistencies in documents using AI"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "autogen-agentchat>=0.4.0",
    "autogen-ext[openai]>=0.4.0",
    "chromadb>=0.5.0",
    "openai>=1.50.0",
    "python-docx>=1.1.0",
    "pymupdf>=1.24.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "tiktoken>=0.7.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "rich>=13.0.0",
    "typer>=0.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[project.scripts]
conflict-detector = "src.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
